additional notification methods
