+++ desktop/src/renderer/pages/transfers/utils/index.ts
@@ -1,2 +1,1 @@
-export * from "./getXlsxFileListToastData";
 export * from "./parseJsonFile";
